Added - Intercom Phone system
What
=================
Added: an Intercom Phone system which would allow players to "Pickup a
phone" on a ground vehicle that has intercom enabled, connecting the
player to the vehicle's intercom system to facilitate communication with
crew or passengers.
This is intended to simulate the hardline phones that are found at the
back of armoured vehicles in the real life military.

Issue ID: N/A