Intercom_Phone: Added - Vehicle EH Changed - Player EH + Loop
What
=================
Added: Vehicle eventhandler in case the vehicle gets destroyed whilst
being connected.
Changed: Player `GetInMan` eventhandler to pass along the vehicle that
is paired with the player, rather than the vehicle the player is getting
into. This means it doesn't matter which vehicle the player gets into,
the intercom phone will disconnect and cleanup appropriately.
Changed: The loop responsible for checking distance of the player from
the vehicle whilst connected. It now uses a more effective CBA function,
whilst also checking for whether the player is still connected to the
intercom phone, so that the loop is exited and cleaned up appropriately
even if the player is still within the correct distance.

Issue ID: N/A