Added - Ext Intercom: Wireless headgear sys & Range/Headgear settings
What
=================
Added: Wireless External Intercom system which works whenever the player
is wearing a wireless capable headgear, defined in `baseHeadgear.hpp`
and/or in the CBA Setting under TFAR Global >> External Intercom. This
is usable with both land and air vehicles.
Added: CBA Settings for the Max Ranges for both Phone and Wireless
systems.

Issue ID: N/A