Added - Ext. Intercom: Eventhandler to automatically keep crew on
What
=================
Added: Eventhandler that will automatically keep any crew who just
dismounted connected to the intercom. The intention is for it to only be
crew, but it's really anyone who is wearing a wireless compatible
headgear and gets out... so doesn't have to be crew.
Actually I may change that to only crew positions in future.

Why
================
To avoid the extra step any crew may have to do to remain connected. Any
extra required interaction, especially under combat, would be
detrimental for the player.

Issue ID: N/A