Changed - Ext Intercom: Headgear whitelisting now done in object
What
=================
Changed: Headgear whitelisting from an array in a global TFAR class, to a property within any headgear item that could be used to whitelist that particular headgear.
The ability to add to this whitelist via a CBA Setting is retained.

Why
=================
To make it easier for mod makers to add compatibility.

Issue ID: N/A